requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.3
cryptography==41.0.7
python-multipart==0.0.6